requirements.txt
python-dotenv
psycopg-binary
cyclonedx-python-lib
